Extract `createSdpOfferEvent` to a separte file
